refactored countries data to be its own file
